fix: update hf list models
