+++ core/providers/huggingface/chat.go
@@ -8,33 +8,6 @@ import (
 	schemas "github.com/maximhq/bifrost/core/schemas"
 )
 
-// ChatCompletion proxies OpenAI-compatible chat completions to Hugging Face inference endpoints.
-func (provider *HuggingFaceProvider) ChatCompletion(ctx context.Context, key schemas.Key, request *schemas.BifrostChatRequest) (*schemas.BifrostChatResponse, *schemas.BifrostError) {
-	if err := providerUtils.CheckOperationAllowed(schemas.HuggingFace, provider.customProviderConfig, schemas.ChatCompletionRequest); err != nil {
-		return nil, err
-	}
-
-	effectiveRequest, resolvedModel := provider.prepareChatRequest(request, key)
-
-	response, bifrostErr := openai.HandleOpenAIChatCompletionRequest(
-		ctx,
-		provider.client,
-		provider.buildRequestURL(ctx, "/chat/completions", schemas.ChatCompletionRequest),
-		effectiveRequest,
-		key,
-		provider.networkConfig.ExtraHeaders,
-		provider.shouldSendRawResponse(ctx),
-		provider.GetProviderKey(),
-		provider.logger,
-	)
-	if bifrostErr != nil {
-		return nil, bifrostErr
-	}
-
-	provider.decorateResponseMetadata(&response.ExtraFields, request.Model, resolvedModel)
-	return response, nil
-}
-
 // ChatCompletionStream forwards streaming chat responses while keeping alias metadata aligned.
 func (provider *HuggingFaceProvider) ChatCompletionStream(ctx context.Context, postHookRunner schemas.PostHookRunner, key schemas.Key, request *schemas.BifrostChatRequest) (chan *schemas.BifrostStream, *schemas.BifrostError) {
 	if err := providerUtils.CheckOperationAllowed(schemas.HuggingFace, provider.customProviderConfig, schemas.ChatCompletionStreamRequest); err != nil {
